State the int -> float conversion lib/draw.cpp's vertex initializers already make.
Draw2DLine and Draw2DRect take int screen coordinates and drop them into a
braced VTX_TL initializer whose x / y / z / rhw are FLOAT. MSVC accepted the
implicit conversion; C++11 forbids narrowing inside braces unless the value is
a constant expression that survives it, so clang refuses exactly the operands
that come from the int parameters and accepts the 0 / 1 / 0.0f / 1.0f beside
them. That is a language-rule refusal, not a missing declaration, which is why
no amount of port/stub/ growth would have reached this TU: it was one of only
two left that needed a game-source edit at all.

#159 re-grounded Hard constraint 1 on behavior preservation and named this
cast as the same class of edit as the path separators and the const on an
operator< that already shipped. It preserves behavior in the strongest sense
available: with -Wno-c++11-narrowing to get the old form past the front end,
clang emits byte-identical assembly before and after at both -O0 and -O2.

x2-1 and y2-1 are parenthesised because (float)x2-1 would move the subtraction
from int into float. The narrowing rule is satisfied either way, so only the
parentheses keep the arithmetic the one the game has always done.

Closes #161
Parent: #124

+++ lib/draw.cpp
@@ -18,7 +18,7 @@
  */
 void Draw2DLine(int x1, int y1, int x2, int y2, D3DCOLOR c1, D3DCOLOR c2){
 	if(c2==0) c2 = c1;
-	VTX_TL vt[] = {x1, y1, 0, 1, c1, x2, y2, 0, 1, c2};
+	VTX_TL vt[] = {(float)x1, (float)y1, 0, 1, c1, (float)x2, (float)y2, 0, 1, c2};
 	//CVertex v;
 	//v.Create(vt, FVF_TL, sizeof(vt));
 	//v.RenderLL();
@@ -53,11 +53,11 @@ void Draw3DLine(VEC3 p1, VEC3 p2, D3DCOLOR c1, D3DCOLOR c2){
  */
 void Draw2DRect(int x1, int y1, int x2, int y2, D3DCOLOR c){
 	VTX_TL vt[] = {
-		x1,		y1,		0.0f, 1.0f, c,
-		x1,		y2-1,	0.0f, 1.0f, c,
-		x2-1,	y2-1,	0.0f, 1.0f, c,
-		x2-1,	y1,		0.0f, 1.0f, c,
-		x1,		y1,		0.0f, 1.0f, c};
+		(float)x1,		(float)y1,		0.0f, 1.0f, c,
+		(float)x1,		(float)(y2-1),	0.0f, 1.0f, c,
+		(float)(x2-1),	(float)(y2-1),	0.0f, 1.0f, c,
+		(float)(x2-1),	(float)y1,		0.0f, 1.0f, c,
+		(float)x1,		(float)y1,		0.0f, 1.0f, c};
 	//CVertex v;
 	//v.Create(vt, FVF_TL, sizeof(vt));
 	//v.RenderLS(4);
